class correction for name input add components for fight
